+++ CHANGES.md
@@ -1,4 +1,11 @@
 
+## 0.3.0 (2021-09-10)
+
+* Move `raster:bands` information in assets (not in properties).
+* update pystac version
+* fix typo for `stddev` raster information
+* drop support of python 3.6 (pystac 1.0.0 dropped support of python 3.6)
+
 ## 0.2.1 (2021-08-24)
 
 * use WarpedVRT for data with internal GCPS